pyproject.toml
[project]
name = "voc-livekit-multimodal-example"
version = "0.1.0"
description = ""
authors = [
    {name = "wira-putra-scoville",email = "wira.putra@sc0ville.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4"
dependencies = [
    "livekit (==0.21.1)",
    "livekit-agents (==0.12.18)",
    "livekit-plugins-azure (==0.5.6)",
    "python-dotenv (==1.0.1)",
    "livekit-plugins-openai (==0.12.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
ruff = "^0.11.4"
types-requests = "^2.32.0.20250328"
poetry-dotenv-plugin = "^0.2.0"
poethepoet = "^0.33.1"
mypy-extensions = "^1.0.0"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = true
indent-style = "space"

[tool.ruff.lint]
select = [
    "A",   # prevent using keywords that clobber python builtins
    "ARG", # flake8-unused-arguments
    "B",   # flake8-bugbear: security warnings
    "C90", # mccabe, previously covered by flake518
    "D",   # pydocstyle
    "E",   # pycodestyle
    "F",   # pyflakes
    "N",   # pep8 naming
    "PL",  # pylint
    "RUF", # ruff rules,
    "SIM", # flake8-simplify
    "UP",  # pyupgrade - very nice tips to adopt sugar syntax if available, great with python 3.10
    "W",   # pycodestyle warnings
    "YTT", # flake8-2020
]
ignore = [
    'D100', 'D101', 'D102', 'D103', 'D104', 'D105', 'D106', 'D107', # pydocstyle errors
    'D205', 'D203', 'D213',
    'D401', 'D406', 'D407', 'D412', 'D413',
    'PLR2004', # pylint magic values
    'RUF012', # avoid conflict with mypy enforcement
    'UP015'
]
preview = true

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.pylint]
max-args = 20
max-locals = 30

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.poe.tasks]
ruff-fix = "ruff check --fix src"
ruff-task = "ruff check src"
mypy-task = "mypy src"
test-task = "coverage run --source=src -m pytest tests/*"
test-report = "coverage report -m"
lint = ["ruff-task", "mypy-task"]
test = ["test-task", "test-report"]
all = ["lint", "test"]

[tool.mypy]
[[tool.mypy.overrides]]
module = "livekit.plugins.turn_detector.*"
ignore_missing_imports = true
